When using authtoken.txt it was found that fgets seems to break things on Linux.
fgets was retaining the newline character causing c_authkey to contain
the line wrap. fgetl was found to work find, but it was opted to use go
ol' fscanf since it just reads the value as a string with no newline
character assumptions. Comments in the source are provided.

Also closed the file upon successful open.